added memory map info from GRUB multiboot

+++ src/printf.c
@@ -24,7 +24,7 @@ int vprintf(void (*write)(char), char const *format, va_list args)
             }
             case 'X':
             {
-                uint32_t arg = va_arg(args, uint32_t);
+                uint64_t arg = va_arg(args, uint64_t);
                 for (int i = 28; i >= 0; i -= 4)
                 {
                     int nibble = (arg >> i) & 0xF;
@@ -36,7 +36,7 @@ int vprintf(void (*write)(char), char const *format, va_list args)
             }
             case 'u':
             {
-                uint32_t arg = va_arg(args, uint32_t);
+                uint64_t arg = va_arg(args, uint64_t);
                 char buf[sizeof arg * __CHAR_BIT__ / 3 + 2]; // Slightly oversized buffer
                 char *buf_end = buf + sizeof buf - 1;
                 char *result = buf_end;
